This currently can't work for most devices
The addition of an example grub.cfg with some more explenation, might be of a lot of help to other people. The current files being generated by this tool is not very useful: the paths are incorrect and there are a few important things missing, additionally you want all your grub files needed to boot in the efi partition; not on a rootfs or it will fail if the first partition dies. Hope this helps, and thanks for your work!